terminando de agregar estilos

+++ src/pages/AdminPanel/ManageUsers.jsx
@@ -1,13 +1,24 @@
 import { useNavigate } from "react-router-dom";
 import useUserCrud from "../../hooks/users/useUserCrud";
-import "./ManageProducts.css";
 import usePagination from "../../hooks/usePagination";
+import { useState } from "react";
+import {
+  Table,
+  Button,
+  Alert,
+  Spinner,
+  Container,
+  Row,
+  Col,
+  Pagination,
+} from "react-bootstrap";
+import "./ManageProducts.css";
 
 export default function ManageUsers() {
   const { users, loading, error, deleteUser } = useUserCrud(
     import.meta.env.VITE_REACT_APP_ROUTE_USUARIOS
   );
-
+  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
   const usersPerPage = 7;
 
   const {
@@ -22,30 +33,56 @@ export default function ManageUsers() {
   const navigate = useNavigate();
 
   const handleDeleteUser = (id) => {
-    deleteUser(id);
-    window.location.reload();
+    const confirmDelete = window.confirm(
+      "¿Seguro que desea eliminar el producto?"
+    );
+    if (confirmDelete) {
+      deleteUser(id);
+      setShowSuccessMessage(true);
+      setTimeout(() => {
+        setShowSuccessMessage(false);
+        window.location.reload();
+      }, 3000);
+    }
   };
 
   const handleEditUser = (id) => {
     navigate(`/edit-user/${id}`);
   };
 
-  if (loading) return <div>Cargando...</div>;
-  if (error) return <div>Error: {error.message}</div>;
+  if (loading)
+    return (
+      <container className="vh-100 vw-100 d-flex justify-content-center align-items-center">
+        <Spinner animation="border" variant="primary" />
+      </container>
+    );
+  if (error)
+    return (
+      <Alert variant="danger">
+        Error: {error.message || "Error desconocido"}
+      </Alert>
+    );
 
   return (
-    <section className="section-manageUsers">
-      <h3>Administrar Usuarios</h3>
-      <div>
-        <h1>usuarios</h1>
-        <button
-          className="table-btn"
-          style={{ marginBottom: "10px", backgroundColor: "#00b894" }}
-          onClick={() => navigate("/create-user")}
-        >
-          Crear Usuario
-        </button>
-        <table className="table">
+    <div className="section-width">
+      <Container className="section-manageUsers">
+        <Row className="my-4">
+          <Col>
+            <h3>Administrar Usuarios</h3>
+          </Col>
+          <Col>
+            <Button variant="success" onClick={() => navigate("/create-user")}>
+              Crear Usuario
+            </Button>
+          </Col>
+        </Row>
+        {showSuccessMessage && (
+          <Alert variant="success" className="toast success-toast">
+            Usuario eliminado con éxito.
+          </Alert>
+        )}
+
+        <Table variant="dark" striped bordered hover responsive>
           <thead>
             <tr>
               <th>Íncice</th>
@@ -66,53 +103,56 @@ export default function ManageUsers() {
                 <td>{user.username}</td>
                 <td>{user.email}</td>
                 <td>{user.esAdmin}</td>
-                <td>
-                  <button
-                    className="table-btn"
+                <td style={{ width: "80px", whiteSpace: "nowrap" }}>
+                  
+                  <Button
+                    variant="primary"
+                    size="sm"
+                    className="me-2"
                     onClick={() => handleEditUser(user._id)}
                   >
-                    Editar
-                  </button>
-                  <button
-                    className="table-btn"
+                    <i className="fi fi-sr-user-pen"></i>
+                  </Button>
+                  <Button
+                    variant="danger"
+                    size="sm"
                     onClick={() => handleDeleteUser(user._id)}
                   >
-                    Eliminar
-                  </button>
+                    <i className="fi fi-ss-trash"></i>
+                  </Button>
                 </td>
               </tr>
             ))}
           </tbody>
-        </table>
+        </Table>
         {/* Paginación */}
-        <div className="pagination-container">
-          <button
-            className="pagination-btn"
-            onClick={handlePreviousPage}
-            disabled={currentPage === 1}
-          >
-            Anterior
-          </button>
-          {Array.from({ length: totalPages }, (_, i) => i + 1).map((page) => (
-            <button
-              key={page}
-              className={`pagination-btn ${
-                page === currentPage ? "active" : ""
-              }`}
-              onClick={() => handlePageChange(page)}
-            >
-              {page}
-            </button>
-          ))}
-          <button
-            className="pagination-btn"
-            onClick={handleNextPage}
-            disabled={currentPage === totalPages}
-          >
-            Siguiente
-          </button>
-        </div>
-      </div>
-    </section>
+        <Row>
+          <Col className="d-flex justify-content-center">
+            <Pagination>
+              <Pagination.Prev
+                onClick={handlePreviousPage}
+                disabled={currentPage === 1}
+              />
+
+              {Array.from({ length: totalPages }, (_, i) => i + 1).map(
+                (page) => (
+                  <Pagination.Item
+                    key={page}
+                    active={page === currentPage}
+                    onClick={() => handlePageChange(page)}
+                  >
+                    {page}
+                  </Pagination.Item>
+                )
+              )}
+              <Pagination.Next
+                onClick={handleNextPage}
+                disabled={currentPage === totalPages}
+              />
+            </Pagination>
+          </Col>
+        </Row>
+      </Container>
+    </div>
   );
 }
